Fixing python script to read json files and moving json commands into the daq folder
