+++ daq/SMELLIE_very_short.json
@@ -42,31 +42,31 @@
 "run_list": [
     {
         "run_id": 1,
-	"hw": [ ],
-	"author": "Chris Jones"
-	"time_stamp": "2010-11-12T13:14:15",
-	"make_smellie_calibration_table": false,
-	"smellie_calibration_table_reference": 1,
-	"trigger_mode": "ORCAtriggersSMELLIE", 		//trigger mode that SMELLIE will use  
+	    "hw": [ ],
+	    "author": "Chris Jones",
+	    "time_stamp": "2010-11-12T13:14:15",
+	    "make_smellie_calibration_table": false,
+	    "smellie_calibration_table_reference": 1,
+	    "trigger_mode": "ORCAtriggersSMELLIE", 		  
         "laser_selected": [2],
         "smellie_driven": true,
-	"naked_laser_intensity": "NULL",
-	"photons_per_pulse": 10000,
+	    "naked_laser_intensity": "NULL",
+	    "photons_per_pulse": 10000,
         "nb_pulses": 10000,
         "smellie_pulse_frequency": 10000
     },
     {
         "run_id": 2,
-	"hw": [ ],
-	"author": "Chris Jones"
-	"time_stamp": "2010-11-12T13:14:15",
-	"make_smellie_calibration_table": true,
-	"smellie_calibration_table_reference": "NULL",
-	"trigger_mode": "ORCAtriggersSMELLIE", 		//trigger mode that SMELLIE will use  
-	"laser_selected": [1],
+	    "hw": [ ],
+	    "author": "Chris Jones",
+	    "time_stamp": "2010-11-12T13:14:15",
+	    "make_smellie_calibration_table": true,
+	    "smellie_calibration_table_reference": "NULL",
+	    "trigger_mode": "ORCAtriggersSMELLIE", 		
+	    "laser_selected": [1],
         "smellie_driven": true,
-	"naked_laser_intensity": 100,
-	"photons_per_pulse": "NULL",
+	    "naked_laser_intensity": 100,
+	    "photons_per_pulse": "NULL",
         "nb_pulses": 100000,
         "smellie_pulse_frequency": 10000
     }
